feat: defaultPlugin can now be a layout plugin
