+++ examples/src/plugins.ts
@@ -34,9 +34,6 @@ import '@react-page/plugins-html5-video/lib/index.css';
 // The image plugin
 import { imagePlugin } from '@react-page/plugins-image';
 import '@react-page/plugins-image/lib/index.css';
-// The rich text area plugin
-import slate from '@react-page/plugins-slate';
-import '@react-page/plugins-slate/lib/index.css';
 // The spacer plugin
 import spacer from '@react-page/plugins-spacer';
 import '@react-page/plugins-spacer/lib/index.css';
@@ -44,12 +41,11 @@ import '@react-page/plugins-spacer/lib/index.css';
 import video from '@react-page/plugins-video';
 import '@react-page/plugins-video/lib/index.css';
 import { ImageUploadType } from '@react-page/ui/lib/ImageUpload/types';
-import React from 'react';
 import customContentPlugin from './customContentPlugin';
 import customContentPluginWithListField from './customContentPluginWithListField';
 import customLayoutPlugin from './customLayoutPlugin';
 import customLayoutPluginWithInitialState from './customLayoutPluginWithInitialState';
-import customSlatePlugin from './customSlatePlugin';
+import { defaultSlate, reducedSlate } from './slate';
 
 const fakeImageUploadService: (url: string) => ImageUploadType = defaultUrl => (
   file,
@@ -71,38 +67,6 @@ const fakeImageUploadService: (url: string) => ImageUploadType = defaultUrl => (
   });
 };
 
-if (
-  process.env.NODE_ENV !== 'production' &&
-  process.env.REACT_APP_TRACE_UPDATES
-) {
-  const { whyDidYouUpdate } = require('why-did-you-update');
-  whyDidYouUpdate(React);
-}
-
-export const defaultSlate = slate(def => ({
-  ...def,
-  plugins: {
-    ...def.plugins,
-    custom: {
-      custom1: customSlatePlugin,
-    },
-  },
-}));
-
-// if you want to use a plugin twice with a differnet config, you have to give it another name
-export const reducedSlate = slate(def => ({
-  ...def,
-  name: def.name + '/reduced', // give it some other name
-  hideInMenu: true, // don't show in insert menu, we only use it as intial children
-  plugins: {
-    headings: {
-      h2: def.plugins.headings.h2,
-      h3: def.plugins.headings.h3,
-    },
-    paragraphs: def.plugins.paragraphs,
-    emphasize: def.plugins.emphasize,
-  },
-}));
 // Define which plugins we want to use.
 
 export const plugins: Plugins = {
