+++ packages/plugins/layout/background/src/types/settings.ts
@@ -1,9 +1,9 @@
-import { BackgroundRendererProps } from './renderer';
-import { BackgroundControlsProps } from './controls';
-import { InitialChildrenDef, ContentPluginConfig } from '@react-page/core';
+import { ContentPluginConfig, InitialChildrenDef } from '@react-page/core';
 import { RGBColor } from '@react-page/ui/lib/ColorPicker/types';
 import { ImageUploadType } from '@react-page/ui/lib/ImageUpload/types';
+import { BackgroundControlsProps } from './controls';
 import { ModeEnum } from './ModeEnum';
+import { BackgroundRendererProps } from './renderer';
 import { Translations } from './translations';
 
 export type BackgroundSettings = {
